backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.4
python-jose[cryptography]==3.4.0
pydantic-settings==2.10.1
firebase-admin==7.1.0
python-multipart==0.0.20
httpx==0.28.1
